scripts/requirements.txt
requests>=2.31.0
python-frontmatter>=1.1.0
markdown>=3.5
markdownify>=0.13.1
